+++ test/transport_test.go
@@ -6,7 +6,7 @@ import (
 	"net/http/httptest"
 	"testing"
 
-	"github.com/wesleyskap/orkai-observability/observability"
+	"github.com/wesleyskap/orkai-observability/v2/observability"
 )
 
 // TestTracingRoundTripperNoActiveTrace asserts that no headers are injected when the LIFO trace stack is empty.